Simplify CLI interface to Template Engine
