--- src/dotnet/commands/dotnet-new3/NullOrEmptyIsLastStringComparer.cs
@@ -1,36 +1,0 @@
-﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System;
-using System.Collections.Generic;
-
-namespace Microsoft.DotNet.Tools.New3
-{
-    internal class NullOrEmptyIsLastStringComparer : IComparer<string>
-    {
-        public int Compare(string x, string y)
-        {
-            if (string.Equals(x, y, StringComparison.OrdinalIgnoreCase))
-            {
-                return 0;
-            }
-
-            if (string.IsNullOrEmpty(x) && string.IsNullOrEmpty(y))
-            {
-                return 0;
-            }
-
-            if (string.IsNullOrEmpty(y))
-            {
-                return 1;
-            }
-
-            if (string.IsNullOrEmpty(x))
-            {
-                return -1;
-            }
-
-            return 0;
-        }
-    }
-}
