--- src/dotnet/commands/dotnet-new3/TableFormatter.cs
@@ -1,84 +1,0 @@
-﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using Microsoft.DotNet.Cli.Utils;
-
-namespace Microsoft.DotNet.Tools.New3
-{
-    internal class TableFormatter
-    {
-        public static void Print<T>(IEnumerable<T> items, string noItemsMessage, string columnPad, char header, Dictionary<string, Func<T, object>> dictionary)
-        {
-            List<string>[] columns = new List<string>[dictionary.Count];
-
-            for (int i = 0; i < dictionary.Count; ++i)
-            {
-                columns[i] = new List<string>();
-            }
-
-            string[] headers = new string[dictionary.Count];
-            int[] columnWidths = new int[dictionary.Count];
-            int valueCount = 0;
-
-            foreach (T item in items)
-            {
-                int index = 0;
-                foreach (KeyValuePair<string, Func<T, object>> act in dictionary)
-                {
-                    headers[index] = act.Key;
-                    columns[index++].Add(act.Value(item)?.ToString() ?? "(null)");
-                }
-                ++valueCount;
-            }
-
-            if (valueCount > 0)
-            {
-                for (int i = 0; i < columns.Length; ++i)
-                {
-                    columnWidths[i] = Math.Max(columns[i].Max(x => x.Length), headers[i].Length);
-                }
-            }
-            else
-            {
-                int index = 0;
-                foreach (KeyValuePair<string, Func<T, object>> act in dictionary)
-                {
-                    headers[index] = act.Key;
-                    columnWidths[index++] = act.Key.Length;
-                }
-            }
-
-            int headerWidth = columnWidths.Sum() + columnPad.Length * (dictionary.Count - 1);
-
-            for (int i = 0; i < headers.Length - 1; ++i)
-            {
-                Reporter.Output.Write(headers[i].PadRight(columnWidths[i]));
-                Reporter.Output.Write(columnPad);
-            }
-
-            Reporter.Output.WriteLine(headers[headers.Length - 1]);
-            Reporter.Output.WriteLine("".PadRight(headerWidth, header));
-
-            for (int i = 0; i < valueCount; ++i)
-            {
-                for (int j = 0; j < columns.Length - 1; ++j)
-                {
-                    Reporter.Output.Write(columns[j][i].PadRight(columnWidths[j]));
-                    Reporter.Output.Write(columnPad);
-                }
-
-                Reporter.Output.WriteLine(columns[headers.Length - 1][i]);
-            }
-
-            if (valueCount == 0)
-            {
-                Reporter.Output.WriteLine(noItemsMessage);
-            }
-
-            Reporter.Output.WriteLine(" ");
-        }
-    }
-}
